backend/requirements.txt
# FastAPI & Server
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
python-multipart>=0.0.6

# Anthropic Claude
anthropic>=0.18.0

# Embeddings
voyageai>=0.2.0
sentence-transformers>=2.3.0

# Vector Database
chromadb>=0.4.22

# Graph
networkx>=3.2

# Data Processing
pandas>=2.2.0
tabulate>=0.9.0

# Document Parsing
pymupdf>=1.23.0
python-pptx>=0.6.23
unstructured[pdf]>=0.12.0

# Configuration
pydantic-settings>=2.1.0
python-dotenv>=1.0.0

# HTTP Client
httpx>=0.26.0

# Testing
pytest>=8.0.0
pytest-asyncio>=0.23.0
